+++ src/util.ts
@@ -1,1 +1,2 @@
+export { figure } from './util/figure';
 export { toc } from './util/toc';