+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add figure syntax.
 - Add label syntax.
+- Add figure function.
 - Refine horizontalrule syntax.
 
 ## 0.42.1